Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { FlatList, View, Text } from 'react-native'
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -15,14 +15,35 @@ class Favorites extends React.Component {
 
   render() {
     const { navigate } = this.props.navigation;
-    
+
 
     const renderMenuItem = ({ item, index }) => {
-      const rightButton = [{
-          text: 'Delete', 
+      const rightButton = [
+        {
+          text: 'Delete',
           type: 'delete',
-          onPress: () => this.props.deleteFavorite(item.id)
-      }];
+          onPress: () => {
+            Alert.alert(
+              'Delete Favorite?',
+              'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+              [
+                {
+                  text: 'Cancel',
+                  onPress: () => console.log(item.name + 'Not Deleted'),
+                  style: ' cancel'
+                },
+                {
+                  text: 'OK',
+                  onPress: () => this.props.deleteFavorite(item.id)
+                }
+              ],
+              { cancelable: false }
+            );
+
+          }
+        }
+      ];
+
       return (
         <Swipeout right={rightButton} autoClose={true}>
           <ListItem
@@ -34,7 +55,7 @@ class Favorites extends React.Component {
             leftAvatar={{ source: { uri: baseUrl + item.image } }}
           />
         </Swipeout>
-        
+
       )
     }
     if (this.props.dishes.isLoading) {
